requirements.txt
django==3.2.14
graphene-django==2.15.0
django-graphql-jwt==0.3.4
